feat(contract-verifier): Add file based config for contract verifier (#2415)
## What ❔
Add file based config for contract verifier
