+++ core/bin/contract-verifier/Cargo.toml
@@ -18,6 +18,7 @@ zksync_contract_verifier_lib.workspace = true
 zksync_queued_job_processor.workspace = true
 zksync_utils.workspace = true
 zksync_vlog.workspace = true
+zksync_core_leftovers.workspace = true
 
 anyhow.workspace = true
 tokio = { workspace = true, features = ["full"] }
