+++ core/lib/zksync_core_leftovers/Cargo.toml
@@ -13,6 +13,8 @@ categories.workspace = true
 zksync_dal.workspace = true
 zksync_config.workspace = true
 zksync_protobuf.workspace = true
+zksync_protobuf_config.workspace = true
+zksync_env_config.workspace = true
 zksync_node_genesis.workspace = true
 
 anyhow.workspace = true
