+++ prover/Cargo.toml
@@ -13,7 +13,6 @@ members = [
     "proof_fri_compressor",
     "prover_cli",
     "prover_version",
-    "config",
 ]
 
 resolver = "2"
@@ -94,10 +93,8 @@ zksync_utils = { path = "../core/lib/utils" }
 zksync_eth_client = { path = "../core/lib/eth_client" }
 zksync_contracts = { path = "../core/lib/contracts" }
 zksync_core_leftovers = { path = "../core/lib/zksync_core_leftovers" }
-zksync_protobuf_config = { path = "../core/lib/protobuf_config" }
 
 # Prover workspace dependencies
-zksync_prover_config = { path = "config" }
 zksync_prover_dal = { path = "prover_dal" }
 zksync_prover_fri_types = { path = "prover_fri_types" }
 zksync_prover_fri_utils = { path = "prover_fri_utils" }
