--- prover/config/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "zksync_prover_config"
-version.workspace = true
-edition.workspace = true
-authors.workspace = true
-homepage.workspace = true
-repository.workspace = true
-license.workspace = true
-keywords.workspace = true
-categories.workspace = true
-
-[dependencies]
-zksync_config.workspace = true
-zksync_env_config.workspace = true
-zksync_core_leftovers.workspace = true
-zksync_protobuf_config.workspace = true
-anyhow.workspace = true
