+++ prover/prover_fri_gateway/Cargo.toml
@@ -15,9 +15,9 @@ zksync_types.workspace = true
 zksync_prover_dal.workspace = true
 zksync_config.workspace = true
 zksync_env_config.workspace = true
+zksync_core_leftovers.workspace = true
 zksync_object_store.workspace = true
 zksync_prover_interface.workspace = true
-zksync_prover_config.workspace = true
 zksync_utils.workspace = true
 zksync_vlog.workspace = true
