+++ prover/prover_fri_gateway/src/main.rs
@@ -4,9 +4,9 @@ use anyhow::Context as _;
 use clap::Parser;
 use reqwest::Client;
 use tokio::sync::{oneshot, watch};
+use zksync_core_leftovers::temp_config_store::{load_database_secrets, load_general_config};
 use zksync_env_config::object_store::ProverObjectStoreConfig;
 use zksync_object_store::ObjectStoreFactory;
-use zksync_prover_config::{load_database_secrets, load_general_config};
 use zksync_prover_dal::{ConnectionPool, Prover};
 use zksync_prover_interface::api::{ProofGenerationDataRequest, SubmitProofRequest};
 use zksync_utils::wait_for_tasks::ManagedTasks;
