+++ prover/witness_generator/Cargo.toml
@@ -14,7 +14,6 @@ vise.workspace = true
 zksync_prover_dal.workspace = true
 zksync_config.workspace = true
 zksync_prover_interface.workspace = true
-zksync_prover_config.workspace = true
 zksync_env_config.workspace = true
 zksync_system_constants.workspace = true
 zksync_vlog.workspace = true
@@ -28,7 +27,6 @@ vk_setup_data_generator_server_fri.workspace = true
 zksync_prover_fri_types.workspace = true
 zksync_prover_fri_utils.workspace = true
 zksync_core_leftovers.workspace = true
-zksync_protobuf_config.workspace = true
 
 zkevm_test_harness = { workspace = true }
 circuit_definitions = { workspace = true, features = [ "log_tracing" ] }
