+++ prover/witness_vector_generator/Cargo.toml
@@ -19,7 +19,7 @@ zksync_object_store.workspace = true
 zksync_prover_fri_utils.workspace = true
 zksync_utils.workspace = true
 zksync_prover_fri_types.workspace = true
-zksync_prover_config.workspace = true
+zksync_core_leftovers.workspace = true
 zksync_queued_job_processor.workspace = true
 zksync_vlog.workspace = true
 vk_setup_data_generator_server_fri.workspace = true
